@@ -1,25 +1,21 @@
 import 'dart:convert';
 
 import 'package:facepunch/lang/l10n.dart';
-import 'package:facepunch/widgets/calendar_strip/date-utils.dart';
+import 'base_model.dart';
+import '/widgets/calendar_strip/date-utils.dart';
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 
 import 'app_const.dart';
 
-class WorkModel with ChangeNotifier{
+class WorkModel extends BaseProvider{
   List<Project> projects = [];
   List<ScheduleTask> tasks = [];
 
   Future<void> getProjectsAndTasks()async{
     try{
-      var res = await http.get(
+      var res = await sendGetRequest(
         AppConst.getProjectsAndTasks,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/x-www-form-urlencoded',
-          'Authorization':'Bearer '+GlobalData.token
-        },
+        GlobalData.token,
       );
       print('[WorkModel.getProjects]${res.body}');
       if(res.statusCode==200){
@@ -38,16 +34,12 @@ class WorkModel with ChangeNotifier{
     }
   }
 
-  Future<String> startShopTracking({String token,int projectId, int taskId})async{
+  Future<String?> startShopTracking({required String token, required int projectId, required int taskId})async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
           AppConst.startShopTracking,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+token
-          },
-          body: {
+          token,
+          {
             'task_id': taskId.toString(),
             'project_id': projectId.toString(),
           }
@@ -66,11 +58,11 @@ class WorkModel with ChangeNotifier{
 }
 
 class Project{
-  int id;
-  int companyId;
-  String name;
-  String companyName;
-  String code;
+  int? id;
+  int? companyId;
+  String? name;
+  String? companyName;
+  String? code;
 
 
   Project.fromJson(Map<String, dynamic> json){
@@ -88,10 +80,10 @@ class Project{
 }
 
 class ScheduleTask{
-  int id;
-  String name;
-  String code;
-  int companyId;
+  int? id;
+  String? name;
+  String? code;
+  int? companyId;
 
   ScheduleTask.fromJson(Map<String, dynamic> json){
     try{
@@ -106,17 +98,17 @@ class ScheduleTask{
 }
 
 class WorkHistory{
-  int id;
-  int userId;
-  int punchId;
-  int taskId;
-  int projectId;
-  String start;
-  String end;
-  String createdAt;
-  String updatedAt;
-  String taskName;
-  String projectName;
+  int? id;
+  int? userId;
+  int? punchId;
+  int? taskId;
+  int? projectId;
+  String? start;
+  String? end;
+  String? createdAt;
+  String? updatedAt;
+  String? taskName;
+  String? projectName;
 
   WorkHistory({
     this.id,
@@ -149,9 +141,10 @@ class WorkHistory{
 
   double workHour(){
     try{
-      if(end==null || end.isEmpty)return 0.0;
-      final startTime = DateTime.parse(start);
-      final endTime = DateTime.parse(end);
+      final startTime = DateTime.tryParse(start!);
+      final endTime = DateTime.tryParse(end!);
+      if(startTime == null)return 0.0;
+      if(endTime == null)return 0.0;
       return endTime.difference(startTime).inMinutes/60;
     }catch(e){
       print('[WorkHistory.workHour]$e');
@@ -159,13 +152,15 @@ class WorkHistory{
     return 0.0;
   }
 
-  DateTime getStartTime(){
-    return DateTime.parse(start);
+  DateTime? getStartTime(){
+    if(start != null){
+      return DateTime.tryParse(start!);
+    }
   }
 
-  DateTime getEndTime(){
-    if(end==null || end.isEmpty) return null;
-    return DateTime.parse(end);
+  DateTime? getEndTime(){
+    if(end==null || end!.isEmpty) return null;
+    return DateTime.tryParse(end!);
   }
 
   String title(){
@@ -189,25 +184,25 @@ class WorkHistory{
   }
 
   bool isEnd(){
-    return end != null && end.isNotEmpty;
+    return end != null && end!.isNotEmpty;
   }
 }
 
-class WorkSchedule{
-  int id;
-  int userId;
-  int projectId;
-  int taskId;
-  String projectName;
-  String taskName;
-  String start;
-  String end;
-  String shift;
-  String color;
-  String noAvailable;
-  bool worked;
-  String createdAt;
-  String updatedAt;
+class WorkSchedule extends BaseModel{
+  int? id;
+  int? userId;
+  int? projectId;
+  int? taskId;
+  String? projectName;
+  String? taskName;
+  String? start;
+  String? end;
+  String? shift;
+  String? color;
+  String? noAvailable;
+  bool? worked;
+  String? createdAt;
+  String? updatedAt;
 
   WorkSchedule({
     this.id,
@@ -266,57 +261,58 @@ class WorkSchedule{
   }
 
   bool isStarted(){
-    return (start !=null && start.isNotEmpty) && (end==null || end.isEmpty);
+    return (start !=null && start!.isNotEmpty) && (end==null || end!.isEmpty);
   }
 
   bool isEnded(){
-    return (start !=null && start.isNotEmpty) && (end !=null && end.isNotEmpty);
+    return (start !=null && start!.isNotEmpty) && (end !=null && end!.isNotEmpty);
   }
 
-  DateTime getStartTime(){
-    return DateTime.parse(start);
+  DateTime? getStartTime(){
+    return DateTime.tryParse(start!);
   }
 
-  DateTime getEndTime(){
-    return DateTime.parse(end);
+  DateTime? getEndTime(){
+    return DateTime.tryParse(end!);
   }
 
   String startTime(){
-    return start.substring(11, 16);
+    if(start == null) return '--:--';
+    return start!.substring(11, 16);
   }
 
   String endTime(){
-    return end.substring(11, 16);
+    if(start == null) return '--:--';
+    return end!.substring(11, 16);
   }
 
-  String isValid(){
-    int startDiff = getStartTime().difference(DateTime.now()).inMinutes;
-    if(startDiff > 5){
-      return 'Why do you start this schedule early?';
-    }else if(startDiff < -5){
-      return 'Why do you start this schedule late?';
+  String? isValid(){
+    final startTime = getStartTime();
+    if(startTime != null){
+      int startDiff = startTime.difference(DateTime.now()).inMinutes;
+      if(startDiff > 5){
+        return 'Why do you start this schedule early?';
+      }else if(startDiff < -5){
+        return 'Why do you start this schedule late?';
+      }
     }
-    int endDiff = getEndTime().difference(DateTime.now()).inMinutes;
-    if(endDiff > 5){
-      return 'Why do you finish this schedule early?';
-    }else if(endDiff < -5){
-      return 'Why do you finish this schedule late?';
+    final endTime = getEndTime();
+    if(endTime != null){
+      int endDiff = endTime.difference(DateTime.now()).inMinutes;
+      if(endDiff > 5){
+        return 'Why do you finish this schedule early?';
+      }else if(endDiff < -5){
+        return 'Why do you finish this schedule late?';
+      }
     }
-    return null;
   }
 
-  Future<String> startSchedule(token)async{
+  Future<String?> startSchedule(token)async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.startSchedule,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/x-www-form-urlencoded',
-          'Authorization':'Bearer '+ token??GlobalData.token
-        },
-        body: {
-          'id' : id.toString()
-        }
+        token??GlobalData.token,
+        { 'id' : id.toString() }
       );
       print('[WorkSchedule.startSchedule]${res.body}');
       if(res.statusCode==200){
@@ -330,16 +326,12 @@ class WorkSchedule{
     }
   }
 
-  Future<String> endSchedule()async{
+  Future<String?> endSchedule()async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.endSchedule,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/x-www-form-urlencoded',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: {'id':id.toString()}
+        GlobalData.token,
+        {'id':id.toString()}
       );
       print('[WorkSchedule.endSchedule]${res.body}');
       if(res.statusCode==200){
@@ -353,16 +345,12 @@ class WorkSchedule{
     }
   }
 
-  Future<String> deleteSchedule()async{
+  Future<String?> deleteSchedule()async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
           AppConst.deleteSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: {'id':id.toString()}
+          GlobalData.token,
+          {'id':id.toString()}
       );
       print('[WorkSchedule.deleteSchedule]${res.body}');
       if(res.statusCode==200){
@@ -376,16 +364,12 @@ class WorkSchedule{
     }
   }
 
-  Future<String> editSchedule()async{
+  Future<String?> editSchedule()async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
           AppConst.editSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: jsonEncode(toJson()),
+          GlobalData.token,
+          toJson(),
       );
       print('[WorkSchedule.editSchedule]${res.body}');
       if(res.statusCode==200){
@@ -399,16 +383,12 @@ class WorkSchedule{
     }
   }
 
-  Future<String> addSchedule()async{
+  Future<String?> addSchedule()async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
           AppConst.addSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: jsonEncode(toJson()),
+          GlobalData.token,
+          toJson()
       );
       print('[WorkSchedule.addSchedule]${res.body}');
       if(res.statusCode==200){
@@ -423,21 +403,21 @@ class WorkSchedule{
   }
 }
 
-class EmployeeCall{
-  int id;
-  int userId;
-  int projectId;
-  int taskId;
-  String projectName;
-  String taskName;
-  String start;
-  String end;
-  String todo;
-  String note;
-  int priority;
-  bool worked;
-  String createdAt;
-  String updatedAt;
+class EmployeeCall extends BaseModel{
+  int? id;
+  int? userId;
+  int? projectId;
+  int? taskId;
+  String? projectName;
+  String? taskName;
+  String? start;
+  String? end;
+  String? todo;
+  String? note;
+  int? priority;
+  bool? worked;
+  String? createdAt;
+  String? updatedAt;
 
   EmployeeCall({
     this.id,
@@ -489,7 +469,7 @@ class EmployeeCall{
       'end':end,
       'todo':todo,
       'note':note,
-      'worked':(worked == null || !worked)?1:0,
+      'worked':(worked == null || !worked!)?1:0,
       'priority':priority,
       'created_at':createdAt,
       'updated_at':updatedAt
@@ -497,39 +477,37 @@ class EmployeeCall{
   }
 
   bool isStarted(){
-    return (start !=null && start.isNotEmpty) && (end==null || end.isEmpty);
+    return (start !=null && start!.isNotEmpty) && (end == null || end!.isEmpty);
   }
 
   bool isEnded(){
-    return (start !=null && start.isNotEmpty) && (end !=null && end.isNotEmpty);
+    return (start !=null && start!.isNotEmpty) && (end !=null && end!.isNotEmpty);
   }
 
-  DateTime getStartTime(){
-    return DateTime.parse(start);
+  DateTime? getStartTime(){
+    return DateTime.tryParse(start??'');
   }
 
-  DateTime getEndTime(){
-    return DateTime.parse(end);
+  DateTime? getEndTime(){
+    return DateTime.tryParse(end??'');
   }
 
   String startTime(){
-    return start.substring(11, 16);
+    if(start == null || start!.length < 16) return '--:--';
+    return start!.substring(11, 16);
   }
 
   String endTime(){
-    return end.substring(11, 16);
+    if(end == null || end!.length < 16) return '--:--';
+    return end!.substring(11, 16);
   }
 
-  Future<String> startCall(String token)async{
+  Future<String?> startCall(String token)async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
           AppConst.startCall,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+token
-          },
-          body: {'id':id.toString()}
+          token,
+          {'id':id.toString()}
       );
       print('[EmployeeCall.startCall]${res.body}');
       if(res.statusCode==200){
@@ -543,16 +521,12 @@ class EmployeeCall{
     }
   }
 
-  Future<String> addEditCall()async{
+  Future<String?> addEditCall()async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.addEditCall,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: jsonEncode(toJson()),
+        GlobalData.token,
+        toJson(),
       );
       print('[EmployeeCall.addEditCall]${res.body}');
       if(res.statusCode==200){
@@ -580,20 +554,21 @@ class EmployeeCall{
       Colors.purple,
       Colors.teal,
     ];
-    return colors[id%11];
+    if(id==null) return Colors.blue;
+    return colors[id!%11];
   }
 }
 
 class EmployeeBreak{
-  int id;
-  int userId;
-  int punchId;
-  String title;
-  String start;
-  int length;
-  bool calculate;
-  String createdAt;
-  String updatedAt;
+  int? id;
+  int? userId;
+  int? punchId;
+  String? title;
+  String? start;
+  int? length;
+  bool? calculate;
+  String? createdAt;
+  String? updatedAt;
 
   EmployeeBreak.fromJson(Map<String, dynamic> json){
     try{
@@ -612,7 +587,10 @@ class EmployeeBreak{
   }
 
   String getTitle(BuildContext context){
-    return '$title ${S.of(context).at} ${PunchDateUtils.getTimeString(DateTime.parse(createdAt))}, $length Minutes';
+    if(createdAt == null){
+      return '$title ${S.of(context).at} --:--, $length Minutes';
+    }
+    return '$title ${S.of(context).at} ${PunchDateUtils.getTimeString(DateTime.tryParse(createdAt!))}, $length Minutes';
   }
 
   Map<String, dynamic> toJson(){
@@ -623,7 +601,7 @@ class EmployeeBreak{
       'start':start,
       'title':title,
       'length':length,
-      'calculate':calculate?1:0,
+      'calculate':(calculate!= null && calculate!)?1:0,
       'created_at':createdAt,
       'updated_at':updatedAt
     };
