@@ -77,7 +77,7 @@ class AppConst{
 }
 
 class GlobalData{
-  static String token;
+  static String token = '';
 }
 
 const int primaryColor = 0xFF09d55b;
